svn merge -r17425:18230 https://eris.llnl.gov/svn/slurm/trunk
And added local mods to support srun --io-timeout

+++ branches/slurm-1.4.srun.hang/testsuite/expect/test1.29.prog.c
@@ -38,20 +38,20 @@ main (int argc, char **argv)
 	int exit_code = 0;
 
 	(void) getrlimit(RLIMIT_CORE, &u_limit);
-	printf("USER_CORE=%d\n", u_limit.rlim_cur);
+	printf("USER_CORE=%d\n", (int)u_limit.rlim_cur);
 	(void) getrlimit(RLIMIT_FSIZE, &u_limit);
-	printf("USER_FSIZE=%d\n", u_limit.rlim_cur);
+	printf("USER_FSIZE=%d\n", (int)u_limit.rlim_cur);
 	(void) getrlimit(RLIMIT_NOFILE, &u_limit);
-	printf("USER_NOFILE=%d\n", u_limit.rlim_cur);
+	printf("USER_NOFILE=%d\n", (int)u_limit.rlim_cur);
 #ifdef RLIMIT_NPROC
 	(void) getrlimit(RLIMIT_NPROC, &u_limit);
-	printf("USER_NPROC=%d\n", u_limit.rlim_cur);
+	printf("USER_NPROC=%d\n", (int)u_limit.rlim_cur);
 #else
 	printf("USER_NPROC unsupported\n");
 #endif
 #ifdef RLIMIT_STACK
         (void) getrlimit(RLIMIT_STACK, &u_limit);
-        printf("USER_STACK=%d\n", u_limit.rlim_cur);
+        printf("USER_STACK=%d\n", (int)u_limit.rlim_cur);
 #else
         printf("USER_STACK unsupported\n");
 #endif
